ASoC: tlv320aic32x4: Set the clock speed in set_dai_sysclk
We are required to have a clock given to us for the driver to run.  Since we
have this master clock, we should set it to the appropriate value when
we are in set_dai_sysclk.

We really should be doing a clk_round_rate to set aic32x4->sysclk in the
function, but changing this to a non-standard value will cause
aic32x4_get_divs to fail later on.