ASoC: tlv320aic32x4: Dynamically calculate PLL values
The PLL values are fixed in a table for different combinations of mclk and
requested sample rate.  This works fine for the precalculated situations.

When mclk is a dynamic clock provided through the common clock framework,
mclk may not be the exact value requested because of rounding and clock
divisor implementations in the hardware.  This will cause the codec driver
to fail to configure itself.  The PLL on the aic32x4 allows for quite a bit
of flexibility to get a good clock rate into the codec if we dynamicaly
calculate its parameters.

This patch dynamically calculates those PLL parameters for a given input
and output frequency.  This should allow a wider variety of clocks to work
with the driver.