i2c: bcm2835: Ensure clock exists when probing
Probe function fails to recognize that upstram clock actually
doesn't yet exist because clock driver has not been initialized.
Actually try to go get the clock and test for its existence
before trying to set up a downstream clock based upon it.

This fixes a bug that causes the i2c driver not to work with
monolithic kernels.

Signed-off-by: Annaliese McDermond <nh6z@nh6z.net>